First Print!
Power Supply need to be 1 for ATX
Y_Min Endstop needed to be inverted.
Z_Min Endstop needed to be inverted.
Y_Max Endstop needed to be inverted.
Z_Max Endstop needed to be inverted.

Final piece…
INVERT_E0_DIR needed to be true to fix the “Extruding-but-really-Retracting/Retracting-but-really-Extruding” thing.